+++ totalRP3/core/impl/popup.lua
@@ -284,15 +284,25 @@ local TRP3_MusicBrowser = TRP3_MusicBrowser;
 local musicWidgetTab = {};
 local filteredMusicList;
 
-local function decorateMusic(lineFrame, musicURL)
-	musicURL = filteredMusicList[musicURL];
-	local musicName = musicURL:reverse();
-	musicName = (musicName:sub(1, musicName:find("%\\")-1)):reverse();
+local function decorateMusic(lineFrame, musicID)
+	local musicName = filteredMusicList[musicID][1];
+	local musicFile = filteredMusicList[musicID][2];
+	local musicDuration = filteredMusicList[musicID][3];
 
-	setTooltipForFrame(lineFrame, lineFrame, "RIGHT", 0, -30, musicName,
-	("|cff00ff00%s\n\n|cffff9900%s: |cffffffff%s\n|cffff9900%s: |cffffffff%s"):format(musicURL, loc.CM_L_CLICK, loc.REG_PLAYER_ABOUT_MUSIC_SELECT2, loc.CM_R_CLICK, loc.REG_PLAYER_ABOUT_MUSIC_LISTEN));
+	local musicShortName = Utils.music.getTitle(musicName);
+	local musicDefaultName = Utils.music.getTitle(musicFile);
+	local tooltipContent;
+	if musicDefaultName == musicShortName then
+		tooltipContent = ("|cff00ff00%s: %ss\n\n|cffff9900%s: |cffffffff%s\n|cffff9900%s: |cffffffff%s"):format(loc.UI_MUSIC_DURATION, floor(musicDuration + 0.5), loc.CM_L_CLICK, loc.REG_PLAYER_ABOUT_MUSIC_SELECT2, loc.CM_R_CLICK, loc.REG_PLAYER_ABOUT_MUSIC_LISTEN);
+	else
+		tooltipContent = ("|cffffff00%s: %s\n|cff00ff00%s: %ss\n\n|cffff9900%s: |cffffffff%s\n|cffff9900%s: |cffffffff%s"):format(loc.UI_MUSIC_ALTTITLE, musicDefaultName, loc.UI_MUSIC_DURATION, floor(musicDuration + 0.5), loc.CM_L_CLICK, loc.REG_PLAYER_ABOUT_MUSIC_SELECT2, loc.CM_R_CLICK, loc.REG_PLAYER_ABOUT_MUSIC_LISTEN);
+		musicName = musicName.."|cffffff00*";
+	end
+
+	setTooltipForFrame(lineFrame, lineFrame, "RIGHT", 0, -30, musicShortName, tooltipContent);
 	_G[lineFrame:GetName().."Text"]:SetText(musicName);
-	lineFrame.musicURL = musicURL;
+
+	lineFrame.musicURL = musicFile;
 end
 
 local function onMusicClick(lineFrame, mousebutton)