+++ totalRP3/modules/chatframe/chatframe.lua
@@ -66,7 +66,6 @@ local CONFIG_OOC_PATTERN = "chat_ooc_pattern";
 local CONFIG_OOC_COLOR = "chat_ooc_color";
 local CONFIG_YELL_NO_EMOTE = "chat_yell_no_emote";
 local CONFIG_INSERT_FULL_RP_NAME = "chat_insert_full_rp_name";
-local CONFIG_INCREASE_CONTRAST = "chat_color_contrast";
 local CONFIG_SHOW_ICON = "chat_show_icon";
 local CONFIG_NPCSPEECH_REPLACEMENT = "chat_npcspeech_replacement";
 
@@ -87,11 +86,6 @@ local function configShowNameCustomColors()
 	return getConfigValue(CONFIG_NAME_COLOR);
 end
 TRP3_API.chat.configShowNameCustomColors = configShowNameCustomColors;
-
-local function configIncreaseNameColorContrast()
-	return getConfigValue(CONFIG_INCREASE_CONTRAST);
-end
-TRP3_API.chat.configIncreaseNameColorContrast = configIncreaseNameColorContrast;
 
 local function configIsChannelUsed(channel)
 	return getConfigValue(CONFIG_USAGE .. channel);
@@ -137,7 +131,6 @@ local function createConfigPage()
 	registerConfigKey(CONFIG_DISABLE_OOC, false);
 	registerConfigKey(CONFIG_REMOVE_REALM, true);
 	registerConfigKey(CONFIG_NAME_COLOR, true);
-	registerConfigKey(CONFIG_INCREASE_CONTRAST, false);
 	registerConfigKey(CONFIG_EMOTE, true);
 	registerConfigKey(CONFIG_EMOTE_PATTERN, "(%*.-%*)");
 	registerConfigKey(CONFIG_OOC, true);
@@ -207,12 +200,6 @@ local function createConfigPage()
 				inherit = "TRP3_ConfigCheck",
 				title = loc.CO_CHAT_MAIN_COLOR,
 				configKey = CONFIG_NAME_COLOR,
-			},
-			{
-				inherit = "TRP3_ConfigCheck",
-				title = loc.CO_CHAT_INCREASE_CONTRAST,
-				configKey = CONFIG_INCREASE_CONTRAST,
-				dependentOnOptions = { CONFIG_NAME_COLOR },
 			},
 			{
 				inherit = "TRP3_ConfigCheck",
@@ -467,7 +454,7 @@ local function wrapNameInColorForNPCEmote(name, senderID, chatColor)
 		if customColor then
 			customColor = Color(petProfile.data.NH);
 
-			if configIncreaseNameColorContrast() then
+			if AddOn_TotalRP3.Configuration.shouldDisplayIncreasedColorContrast() then
 				customColor:LightenColorUntilItIsReadableOnDarkBackgrounds();
 			end
 
@@ -631,7 +618,6 @@ TRP3_API.chat.getFullnameForUnitUsingChatMethod = getFullnameForUnitUsingChatMet
 local defaultGetColoredNameFunction = GetColoredName;
 
 local GetClassColorByGUID = TRP3_API.utils.color.GetClassColorByGUID;
-local GetCustomColorByGUID = TRP3_API.utils.color.GetCustomColorByGUID;
 
 -- This is our custom GetColoredName function that will replace player's names with their full RP names
 -- and use their custom colors.
@@ -676,6 +662,8 @@ function Utils.customGetColoredNameWithCustomFallbackFunction(fallback, event, a
 	end
 	-- Make sure we have a unitID formatted as "Player-Realm"
 	unitID = unitInfoToID(character, realm);
+	---@type Player
+	local player = AddOn_TotalRP3.Player.static.CreateFromGUID(GUID)
 
 	-- Character name is without the server name is they are from the same realm or if the option to remove realm info is enabled
 	if realm == Globals.player_realm_id or getConfigValue(CONFIG_REMOVE_REALM) then
@@ -694,15 +682,7 @@ function Utils.customGetColoredNameWithCustomFallbackFunction(fallback, event, a
 	end
 
 	if configShowNameCustomColors() then
-		local customColor = GetCustomColorByGUID(GUID);
-
-		if customColor then
-			if configIncreaseNameColorContrast() then
-				customColor:LightenColorUntilItIsReadable();
-			end
-
-			characterColor = customColor;
-		end
+		characterColor = player:GetCustomColorForDisplay() or characterColor;
 	end
 
 	-- If we did get a color wrap the name inside the color code